Add containers molecule driver
As this image is superseding the ansible toolset image, it should aim at
feature parity with its precursor.

This change adds the molecule containers driver - which is a drop-in
replacement for the existing Docker and Podman drivers - and its necessary
dependencies. This also includes the community.docker and containers.podman
ansible collections, as written in the molecule-docker and molecule-podman
README.

See: ansible#20

Signed-off-by: Daniel Ziegenberg <daniel@ziegenberg.at>