add podman and docker binaries
In #22 and #23 the podman and docker molecule drivers were added,
but the respective binaries are still missing.

This commit adds the necesarry dependencies to the execution
enviroments bindep.txt.

This commit also adds some CLI commands to check we do not remove
them inadvertently and output some helpful version information at
build time.

There is also a workaround to make installing podman and docker
simultaneously possible. This is due to a conflict between the
packaged version of runc in the containerd.io package from Docker
and the CentOS 8 Stream native version packaged for Podman and Skopeo.
This can be changed once https://github.com/docker/containerd-packaging/pull/231
is merged and available upstream via the Docker repository.
Cudos for the workaround:
https://faun.pub/how-to-install-simultaneously-docker-and-podman-on-rhel-8-centos-8-cb67412f321e

Signed-off-by: Daniel Ziegenberg <daniel@ziegenberg.at>