Remove molecule-docker and containers plugins
This change is needed because community.docker collections has
a requirements on python package docker-compose<2.0 and that is
causing other conflicts with our dependencies. As new docker-compose
versions are no longer on pypi, the only workaround would for the
collection to make that dependency soft.

This will not be added back unless that conflicts are addressed.